⚡️Change uint256 --> uint

+++ contracts/PlearnERC20.sol
@@ -9,14 +9,14 @@ contract PlearnERC20 is IPlearnERC20 {
     string public constant override name = 'Plearn LPs';
     string public constant override symbol = 'PLN-LP';
     uint8 public constant override decimals = 18;
-    uint256 public override totalSupply;
-    mapping(address => uint256) public override balanceOf;
-    mapping(address => mapping(address => uint256)) public override allowance;
+    uint public override totalSupply;
+    mapping(address => uint) public override balanceOf;
+    mapping(address => mapping(address => uint)) public override allowance;
 
     bytes32 public override DOMAIN_SEPARATOR;
     // keccak256("Permit(address owner,address spender,uint256 value,uint256 nonce,uint256 deadline)");
     bytes32 public constant override PERMIT_TYPEHASH = 0x6e71edae12b1b97f4d1f60370fef10105fa2faae0126114a169c64845d6126c9;
-    mapping(address => uint256) public override nonces;
+    mapping(address => uint) public override nonces;
 
     constructor() {
         DOMAIN_SEPARATOR = keccak256(
